requirements_prod.txt
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorclass==2.2.2
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.0.0
Django==3.2.8
django-auditlog==1.0.0
django-celery-beat==2.2.1
django-cors-headers==3.13.0
django-jsonfield-backport==1.0.4
django-on-heroku==1.1.2
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docopt==0.6.2
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.4
MarkupSafe==2.1.0
packaging==21.3
paystackapi==2.1.0
pip-upgrader==1.4.15
prompt-toolkit==3.0.31
psycopg2==2.9.3
PyJWT==2.4.0
pyotp==2.7.0
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==3.5.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
terminaltables==3.1.10
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
